==> docs/01_assign_catchments.Rmd <==
---
title: "Reassigning Catchments"
output: 
    html_document:
        toc: true
        toc_float: true
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = FALSE, warning=FALSE, message=FALSE)

# libraries
library(sf)
library(tidyverse)
library(mapview)
library(tmap)
library(tmaptools)
library(glue)
library(mapedit)
library(leafpm)
mapviewOptions(fgb=FALSE)

```

As part of the CEFF case study of the Little Shasta, we identified a number of inaccuracies associated with the NHD watershed and streamline delineation.
In part, this was because canals have been treated as streamlines.
In particular, the canal which delivers water from Lake Shastina is routed as if it is a drainage to the Little Shasta.

## Recalculating Drainage Area & Streamlines

To address inaccuracies in the NHD flowlines and catchments associated with designation of canals as streams, the following steps were taken:

1.  We identified natural stream channel from artificial (canal) channels, and filtered all artificial segments out, and fixed mis-classified segments.
2.  Identify catchments based on the new clean streamline dataset, and create a revised catchment layer. We attributed each catchment to a mainstem flowline on the Little Shasta.
3.  Calculate the total drainage area for each mainstem NHD COMID segment and add this attribute information as a new column in the stream layer. Retain the old drainage area associated with each NHD segment to facilitate comparisons.
4.  Calculate the difference between the two drainage areas, and adjust **magnitude** functional flow metrics by this percentage difference (e.g., if the adjusted drainage area is 20% smaller than the original value, the magnitude metrics are reduced by 20%).

**Magnitude Metrics:**

-   `DS_Mag_50, DS_Mag_90, FA_Mag, Peak_10, Peak_5, Peak_2, SP_Mag, Wet_BFL_Mag_10, Wet_BFL_Mag_50`

## Basic Map

```{r getData}

load("data_output/little_shasta_catchment_flowlines.rda")

# shrink the data down
flowlines_map <- flowlines %>% select(id, comid, hydroseq, gnis_name, areasqkm:divdasqkm, shape_length, streamorde, streamorder_map, streamcalc)

# map
m1 <- mapview(catch_h10, alpha.regions=0.1, color="maroon", lwd=0.5,legend=FALSE) +
    mapview(evans, layer.name="Evans Streamline", color="darkblue") +
    #mapview(lsh_springs,layer.name="Springs", col.regions="cyan4") +
    #mapview(h10, layer.name="HUC10", alpha.regions=0, color="gray20", lwd=2)+
    mapview(flowlines_map, color="dodgerblue", legend=FALSE)

```

### Select Features

Select catchments for each mainstem stream seg starting downstream.

```{r selectMap, echo=T, eval=F}

#comids
# 3917200, 3917198, 3917948, 3917950, 3917244, 3917946
comid_s <- 3917950

# run and draw whatever shape or line through thingg
df_sel <- selectFeatures(catch_h10, map = m1@map, mode = "draw")

# tidy
df_sel <- df_sel %>% 
  select(-edit_id)

# add a comid
df1 <- df_sel %>% 
    mutate(comid=comid_s)

# check with:
mapview(df1, zcol="FEATUREID") + mapview(flowlines_map, color="dodgerblue", legend=F) + mapview(catch_h10, color="orange", col.regions=NA, legend=F, lwd=0.2)

write_rds(df1, file = glue("data_output/ls_{comid_s}_S.rds"))
write_rds(df1, file = glue("data_output/ls_{comid_s}_N.rds"))

```


## Look at Catchments

```{r}

library(fs)
library(purrr)
catchs <- fs::dir_ls(path = "data_output", type = "file", regexp = "ls_[0-9]*?_")

catchments <- map(catchs, ~read_rds(.x))
df_catch <- bind_rows(catchments) %>% st_transform(3310)
# fix comid issue *39171948 should be 3917948
df_catch <- df_catch %>% mutate(comid = case_when(
  grepl(39171948, comid) ~ 3917948,
  TRUE ~ comid
))

df_catch$comid_f <- as.factor(df_catch$comid)
flowlines_map$comid_f <- as.factor(flowlines_map$comid)

# filter catchs from upper watershed:
catch_h10_us <- catch_h10 %>% filter(!FEATUREID %in% tst$FEATUREID)
catch_h10_us$comid <- catch_h10_us$FEATUREID
catch_h10_us$upper <- TRUE

# bind together into one, fix crs first
st_crs(df_catch)==st_crs(catch_h10_us)
catch_final <- bind_rows(df_catch, catch_h10_us)
st_crs(catch_final)

# make comid a factor
catch_final$comid_f <- as.factor(catch_final$comid)


mapview(df_catch, zcol="comid_f", col.regions=colorspace::qualitative_hcl(6, palette = "dark2"), legend=TRUE, lwd=0.6, layer.name="COMID") + 
  mapview(flowlines_map, color="blue2", legend=F) + 
  mapview(catch_h10_us, color="black", alpha.col=0.8, col.regions=NA, legend=FALSE, lwd=0.6) + 
  mapview(evans, layer.name="Evans Streamline", color="darkblue") +
  mapview(lsh_springs,layer.name="Springs", col.regions="cyan4") +
  mapview(h10, layer.name="HUC10",col.regions="gray20", alpha.col=0.5, alpha.regions=0, color="gray20", lwd=3)

```


## Make Static Map

```{r}

library(tmap)
library(tmaptools)

load("data_output/tmaptools_h10_osm_natgeo.rda")

# make two versions of flowlines: mainstem and otherwise
flowlines_main <- flowlines %>% filter(streamorde>=3)
flowlines_tribs <- flowlines %>% filter(streamorde<3)

# map
(map_base <-
    tm_shape(gm_osm) + tm_rgb() +
    tm_shape(df_catch) +
    tm_fill(col = "comid_f", alpha = 0.5, title = "Catchment COMID") +
    tm_shape(catch_h10) +
    tm_polygons(border.col="gray50", alpha = 0, border.alpha = 0.7, lwd=0.7, lty=2) +
    tm_shape(h10) +
    tm_polygons(border.col="gray30", alpha = 0, border.alpha = 0.9, lwd=4) +
    tm_shape(flowlines_map) + tm_lines(col="darkblue", lwd="streamcalc", scale = 2.25, legend.lwd.show = FALSE) +
    tm_shape(evans) + tm_lines(col="darkblue", lwd=0.5) +
    tm_shape(lsh_springs) +
    tm_dots(col="skyblue", size = 1.2, title = "Springs", legend.show = TRUE, shape = 21) +
    tm_text("Name", auto.placement = 0.7, just = "left", xmod = 0.35, shadow = TRUE)+
    tm_layout(frame=FALSE) +
    tm_layout(title = "Little Shasta",
              frame = FALSE,
              fontfamily = "Roboto Condensed",
              legend.outside = FALSE, attr.outside = FALSE,
              inner.margins = 0.01, outer.margins = (0.01),
              #legend.position = c(0.6,0.85),
              title.position = c(0.7, 0.95)) +
    tm_compass(type = "4star", position = c("left","bottom")) +
    tm_scale_bar(position = c("left","bottom")))

# save
tmap_save(map_base, filename = "figs/map_of_catchments_assigned.jpg", height = 8.5, width = 11, units = "in", dpi = 300)

```


## To Do

 - is model stream based or catchment based? do we need a streamline to assign FFM to watershed, or is it catchment based

 - Step I, identify ecological goals, locations of interest to tag objectives, don't have to be tied to a gage (ideal if possible, but not necessary). Could be areas in the watershed you want to focus discussion and focus...could be where spring comes in.
