==> docs/05_network_for_little_shasta.Rmd <==
---
title: "Little Shasta Stream Network"
author: "Ryan Peek"
date: "Updated: `r format(Sys.Date())`"
output: html_document
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r}

# Libraries
suppressPackageStartupMessages({
  library(tidyverse);
  library(nhdplusTools);
  library(glue);
  library(sf);
  library(stringr);
  library(igraph);
  library(networkD3);
  library(here);
  library(mapview)})
mapviewOptions(fgb = FALSE)

```

## Little Shasta Data

Here we are using an original flowlines layer that is all streamlines in the watershed. It includes 140 variables associated with each COMID from the NHDPlus dataset. This also includes a few additional pieces including: 

 - `evans` a streamline hand digitized from a spring. 
 - `gages` the 3 USGS/UCD gages in the watershed
 - `h10_ls` the HUC10 boundary revised to include/match subcatchment boundaries
 - `lsh_springs` a list of the main springs that are relevant to functional flow analysis

```{r data}

load(here("data_output/05_network_data_pieces.rda"))

```

Clean up and identify the Locations of Interest (LOI) which are COMIDs in the lower half of the watershed.

```{r}

# reduce fields for plotting purposes
flowlines_map <- flowlines %>% select(id, comid, contains("seq"), hydroseq, gnis_name, areasqkm:divdasqkm, shape_length, streamorde, streamorder_map, streamcalc)

# fix flowlines comid to factor in correct order
flowlines_map <- flowlines_map %>%
  mutate(comid_f = factor(as.character(comid),
                          levels=c(3917198, 3917200, 3917948,
                                   3917950, 3917244, 3917946)))


```


## Overview Map

We've cleaned and removed the canals/non-streams from the catchment map, so this is the overview. Notice there are some "sinks" that do not drain to the mainstem.

```{r mapoverview, echo=FALSE}

mapview(loi_comid, color="coral1", lwd=4, layer.name="LOI Comids") +
  mapview(flowlines_map, color="darkblue", legend=F, lwd=2) +
  mapview(gages, col.regions="black", color="white", cex=5, layer.name="Gages") +
  mapview(catch_final, color=scales::alpha("black",0.2), alpha.regions=0.2, col.regions=NA, legend=FALSE, lwd=0.2) +
  mapview(lsh_springs,layer.name="Springs", col.regions="cyan4") +
  mapview(h10_ls, col.regions="steelblue", alpha.regions=0, color="steelblue", lwd=3,
           layer.name="HUC10 Revised", legend=TRUE)


```

## Leverage the NLDI for Network Flowline Data

The {`nhdplusTools`} package is great for accessing NHDPlus data and attributes. Here we can test it out to grab flowlines upstream of a given point or COMID. 

The flow network is essentially correct starting near the most upstream LOI (COMID==3917948), so we can use that to grab flowlines and test things out. There are multiple sources that can be used to query the NLDI (see `dataRetrieval::get_nldi_sources()`). We will use COMID.

First we make the nldi call for our specific comid.

```{r}
# COMID above diversion: 3917948
# COMID at confluence w Shasta River: 3917946

# get list of comids:
nld_ls <- list(featureSource = "comid",
               featureID = 3917948)
```

Next we can use this to navigate upstream or downstream and pull flowlines, attributes, nodes, etc. Here we will pull all flowlines upstream.

```{r}
# now navigate and plot (UPSTREAM)
# here we get all mainstem and tributary branches for 120 km upstream
get_us <- navigate_nldi(nldi_feature = nld_ls,
              mode = "upstreamTributaries",
              distance_km = 120)$UT 

mapview(get_us, zcol="nhdplus_comid", legend=FALSE) +
  mapview(flowlines_map, color="steelblue", lwd=0.3)
```

And here we can pull all mainstems only (so no diversions or canals), "*DM*" stands for *downstream mainstem*.

```{r}
# now get DOWNSTREAM
get_dm <- navigate_nldi(nldi_feature = nld_ls,
                           mode = "DM",
                           distance_km = 15)$DM #%>%

```

### Plot the Flowline Map

Bind it all together (to get hydroseq data) and plot. 

```{r}

# bind the updated flowlines
nhd_flow <- bind_rows(get_dm, get_us)

df <- flowlines_map %>% filter(comid %in% nhd_flow$nhdplus_comid)

```

Map it!

```{r, echo=FALSE}

mapview(nhd_flow, zcol="nhdplus_comid", legend=FALSE, layer.name="NHD Flowlines", lwd=4) +
  mapview(df, zcol="hydroseq", layer.name="Hydroseq", lwd=2)

```

### Test with An Upstream COMID

Here we can test this with a different comid on one of the smaller tributaries, should pull just those comids.

```{r specNhdcomid}
nld_ls_seg <- list(featureSource = "comid",
               featureID = 3917920)

get_us_seg <- navigate_nldi(nldi_feature = nld_ls_seg,
              mode = "upstreamTributaries",
              distance_km = 120)$UT 

mapview(get_us_seg)

# the comids
get_us_seg$nhdplus_comid

```

So we could just iterate through this NHD call, but that seems inefficient to some degree.



# iGraph Network

To use this in a watershed, replace the `up_down_df` below with your two-column adjacency/edge dataframe.

Be sure that column headings match (`up`, `down`) or change the names of column headings throughout this script to match your input dataframe.

## Get Flowline Data

```{r getdata}

# set up two-column edge dataframe
flowlines <- read_rds(here("data_output/final_flowlines_w_full_nhd_vaa.rds"))

# reduce fields for plotting purposes
flowlines_map <- flowlines %>% select(id, comid, contains("seq"), hydroseq, gnis_name, areasqkm:divdasqkm, shape_length, streamorde, streamorder_map, streamcalc, geom)

# fix flowlines comid to factor in correct order
flowlines_map <- flowlines_map %>%
  mutate(comid_f = factor(as.character(comid),
                          levels=c(3917198, 3917200, 3917948,
                                   3917950, 3917244, 3917946)))

# drop sinks (isolated channels)
sinks <- c(3917228, 3917212, 3917214, 3917218, 3917220,
           3917960, 3917958, 3917276, 3917278, 3917274, 
           3917282, 3917284, 3917286, 3917280, 3917268,
           3917256, 3917250, 3917272, 3917956)

flowlines_map_trim <- flowlines_map %>% 
  filter(!comid %in% sinks)

mapview(flowlines_map_trim)

```

## Create adjacency matrix

Columns contain vertices and each full row represents an edge (from a vertex to vertex). 

For a watershed, this will be a *to-from* dataframe with hydroseq values associated with nodes. We can use NAs to flag terminal values for ease of comparison with network image, but can be filtered out to create the igraph. This drops vertices that do not compose edges.

```{r getadj-ex, echo=FALSE, eval=FALSE}
# demo
up_down_df <- tribble(
  ~up, ~down,
  "Q", NA,
  "P", "N",
  "O", "N",
  "N", "M",
  "M", NA,
  "L", "G",
  "K", "G",
  "J", "G",
  "I", "E",
  "H", "E",
  "G", "C",
  "F", "B",
  "E", "B",
  "D", "A",
  "C", "A",
  "B", "A",
  "A", NA,
  "R", "Q",
  NA, "P",
  NA, "O",
  NA, "L",
  NA, "K",
  NA, "J",
  NA, "I",
  NA, "H",
  NA, "F",
  NA, "D",
  NA, "R",
  "S", "T",
  "T", NA,
  "R", "S",
  "U", NA,
  NA, "U"
  ) %>%
  filter(!is.na(up) & !is.na(down))

# tst
(p <- simpleNetwork(up_down_df, height = "400px", 
                    width = "400px", 
                    fontSize = 16, 
                    fontFamily = "serif",
                    nodeColour = "darkblue", 
                    linkColour = "steelblue",
                    opacity = 0.9, zoom = TRUE, charge = -40))
```

### Little Shasta Network

Here we can create our network.

```{r getadj-lsh}

library(nhdplusTools)

# get nodes: Start
starts <- get_node(flowlines_map_trim, "start") %>%
  mutate(hydroseq=flowlines_map_trim$hydroseq) %>% 
  arrange(hydroseq) %>% 
  mutate(us_ord = rev(seq(1:nrow(.))))

# get midpoints
midpts <- st_point_on_surface(flowlines_map_trim)

mapview(starts, zcol="hydroseq", legend=FALSE,
        layer.name="Starts") + 
  mapview(flowlines_map_trim, legend=FALSE) +
  mapview(flowlines_map) +
  mapview(midpts, col.regions="green", cex=3)

# get matrix
up_down_lsh <-
  midpts %>% st_drop_geometry() %>% 
  select(up = hydroseq,
         down = dnhydroseq) %>%
  # fix the downstream end: 10020426 to NA or zero?
  # mutate(down = case_when(
  #   up == 10020426 ~ 1, # this is OUTLET into Shasta: 10019615
  #   TRUE ~ down)) %>%
  filter(!is.na(down)) # make sure no NAs

# spur at ds end is: 10135201, penultimate seg: 10020862

# tst
(p <- simpleNetwork(up_down_lsh, Source = "up", Target = "down", height = "400px", 
                    width = "400px", 
                    fontSize = 16, 
                    fontFamily = "serif",
                    nodeColour = "darkblue", 
                    linkColour = "steelblue",
                    opacity = 0.9, zoom = TRUE, charge = -100))

```

## Identify Termini

Need to update and identify the termini to create the "edges" of the network.

### Upstream Ends

We want to identify the upstream ends, which is where paths stop. To run the other direction (upstream --> downstream), we can change what is considered "upstream" or reverse the `all_simple_paths` call below.

This way of doing it uses an edge data frame, but we could also specify upstream ends if needed, and they don't have to be termini (e.g. you can stop mid-network).

```{r usends}

# this identifies the upstream "ends" (termini)
upstream_ends <- up_down_lsh %>%
  filter(!up %in% down) %>%
  pull(up) %>%
  unique()
```

### Downstream Ends

Next we can identify downstream ends. The upstream ends are the places paths stop, so to run the other direction (upstream --> downstream), we can change  what is considered "downstream" or reverse the `all_simple_paths` call below. 

```{r dsends}

downstream_ends <- up_down_lsh %>%
  filter(!down %in% up) %>%
  pull(down) %>%
  unique()

```

## Convert to Matrix

```{r tomatrix, fig.cap="An idealized network for Little Shasta, the light blue is the downstream outlet, the dark blue is the lower bottomlands mainstem."}

# turn edge dataframe into matrix for igraph conversion
up_down_matrix <- up_down_lsh %>% 
  mutate(across(everything(), as.character)) %>% 
  as.matrix()

# convert edge matrix to igraph
# note that directionality isn't attributed in the igraph object
# directionality is implicit when paths are created using upstream and downstream ends
# creating an igraph object without directionality allows for upstream path creation
test_igraph <- graph_from_edgelist(up_down_matrix, directed = FALSE)

# double check visually
mainstems <- list(c("10019615", "10020426", "10020862", "10021351","10032346", "10033986","10035874","10038060","10040642","10043744", "10047601","10052559","10059365"), c("10019615"))
mainstems_end <- list(c("10019615", "10020426", "10020862", "10021351","10032346"),c("10019615"))
#plot(test_igraph, mark.groups = mainstems, mark.col = c("steelblue","cyan4"), mark.shape = .4) # this breaks/hangs if matrix numeric
plot(test_igraph, mark.groups = mainstems_end, mark.col = c("steelblue","cyan4"), mark.shape = 0.8) # this breaks/hangs if matrix numeric
```

### Convert to a Dataframe

```{r functions}
# compute all simple paths between downstream ends and upstream ends
# and store them in a list of igraph paths
hold_paths <- lapply(as.character(downstream_ends),
                     function(x) all_simple_paths(test_igraph,
                                                  from = x,
                                                  to = as.character(upstream_ends)))

# unplist igraph paths
hold_paths <- unlist(hold_paths, recursive = F)

# drop unused information like geodesic attributes from igraph paths
paths <- lapply(1:length(hold_paths), function(x) as_ids(hold_paths[[x]]))

# turn list of paths into dataframe
paths_df <- as.data.frame(as.matrix(paths)) %>%
  magrittr::set_colnames("paths_lst") %>%
  rowid_to_column(var = "row_id")

# create final dataframe of paths that has
# changing the `collapse` argument will change the separator in the path strings
paths_df_clean <- unnest(paths_df, paths_lst) %>%
  group_by(row_id) %>%
  mutate(paths_chr = paste0(paths_lst, collapse = " ")) %>%
  select(row_id, paths_chr) %>%
  unique() %>%
  mutate(downstream_end = str_sub(paths_chr, 1, 1),
         upstream_end = str_sub(paths_chr, -1, -1)) %>%
  left_join(paths_df, by = "row_id") %>% 
  select(row_id, downstream_end, upstream_end, paths_chr, paths_lst)
  
```

### Look at the Data

```{r cleandf}

# add most upstream id
paths_df_clean$us_hydroid <- unlist(map_depth(paths_df_clean$paths_lst, 1, ~tail(.x, 1)))

# take a look at the output dataframe
DT::datatable(paths_df_clean)

# write_csv(paths_df_clean, file = here("data_output/lshasta_hydroid_network.csv"))
```


## to do

 - make sure comid/hydroid assigned to each catchment correlates with the flowline network we want.
 - ignore the isolated drainages and assign to the places we want them to join
 - follow up and review the X-walk to see what variables we are missing
 - create google sheet with Xwalk so collaboration w Ted can happen.
 - add a column with potential calculation needed (i.e., mean, dominant, etc)
 - start drafting code to run accumulation
